Refactor so template based formatting is easier.
